updated db

+++ src/MyConnection.java
@@ -1,7 +1,9 @@
 import java.sql.*;
+import org.apache.log4j.Logger;
 
 public class MyConnection {
 
+    private static final Logger log = Logger.getLogger(MyConnection.class);
     static Connection conn = null;
     static Statement statement = null;
 
@@ -43,7 +45,10 @@ public class MyConnection {
     public static synchronized String insertIntoDB(String ldap_login, String rep_fam,
                                                    String rep_name, String rep_ot,
                                                    Date rep_birth, String rep_posit){
+        log.info("Начало обработки строки");
         if(checkLoginInDB(ldap_login)) return "Already exists";
+
+        if(ldap_login.length() > 12) return "Too long login";
 
         String selectTableSQL = "insert into test.rep_emp values('"
                 + ldap_login + "', '"
@@ -52,24 +57,25 @@ public class MyConnection {
                 + rep_ot  + "', '"
                 + rep_birth  + "', '"
                 + rep_posit  + "')";
-               // + "', '', '', '', '" + date + "', '');";
-        System.out.println(selectTableSQL);
+
         try {
-            //длина не может быть больше 12
             statement.executeUpdate(selectTableSQL);
         } catch (SQLException e) {
-            //log
+            //log System.out.println(selectTableSQL);
+
+            log.info("Еррок");
             e.printStackTrace();
             return "Error";
         }
 
+        log.info("Начало обработки строки");
         return "Success";
     }
 
     private static boolean checkLoginInDB(String login){
 
         try {
-            String selectTableSQL = "select ldap_login from test.rep_emp where ldap_login = '"
+            String selectTableSQL = "select distinct ldap_login from test.rep_emp where ldap_login = '"
                                     + login + "';";
             ResultSet rs = statement.executeQuery(selectTableSQL);
             System.out.println(selectTableSQL);
@@ -82,25 +88,43 @@ public class MyConnection {
         return false;
     }
 
-    public static synchronized void deleteFromDB(String str){
+    public static synchronized String deleteFromDB(String ldap_login){
+
+        if(!checkLoginInDB(ldap_login)) return "Deleted";
 
             String selectTableSQL = "delete from test.rep_emp"
-                   + " where test.rep_emp.LDAP_LOGIN = '" + str + "'";
+                   + " where test.rep_emp.LDAP_LOGIN = '" + ldap_login + "'";
 
-        System.out.println(selectTableSQL);
+
         try {
             statement.execute(selectTableSQL);
         } catch (SQLException e) {
-            //log
+            //log System.out.println(selectTableSQL);
             e.printStackTrace();
+            return "Error";
         }
+
+        return "Success";
     }
 
-    public static synchronized void updateInDB(){
+    public static synchronized String updateInDB(String ldap_login,
+                                                 String attr_name,
+                                                 String attr_value){
+
+        if(!checkLoginInDB(ldap_login)) return "No such user";
+
         String selectTableSQL = "update test.rep_emp"
-        + " set test.rep_emp.REP_POSIT = '224488'"
-        + " where LDAP_LOGIN = 'DN118899DAG2'";
-        System.out.println(selectTableSQL);
+        + " set test.rep_emp." + attr_name + " = '" + attr_value + "'"
+        + " where LDAP_LOGIN = '" + ldap_login + "'";
+
+        try {
+            statement.execute(selectTableSQL);
+        } catch (SQLException e) {
+            System.out.println(selectTableSQL);
+            e.printStackTrace();
+            return "Error";
+        }
+        return "Success";
     }
 
     private static void createDbUserTable(Connection con) throws SQLException {
